+++ android-sdk-ui/src/main/assets/braze-html-bridge.js
@@ -36,6 +36,7 @@ var brazeBridge = {
     setDateOfBirth: function(year, month, day) { brazeInternalBridge.getUser().setDateOfBirth(year, month, day); },
     setCountry: function(country) { brazeInternalBridge.getUser().setCountry(country); },
     setPhoneNumber: function(phone) { brazeInternalBridge.getUser().setPhoneNumber(phone); },
+    setLineId: function(lineId) { brazeInternalBridge.getUser().setLineId(lineId); },
     setCustomUserAttribute: function(key, value, merge = false) {
       var isArray = function(value) {
           if (Array.isArray) {
@@ -68,7 +69,8 @@ var brazeBridge = {
   web: {
     registerAppboyPushMessages: function(successCallback, deniedCallback) { console.log("This method is a no-op on Android."); },
     trackLocation: function() { console.log("This method is a no-op on Android."); },
-  }
+  },
+  setBannerHeight: function(height) { brazeInternalBridge.setBannerHeight(height); }
 };
 var appboyBridge = brazeBridge;
 window.dispatchEvent(new Event("ab.BridgeReady"));