+++ android-sdk-ui/src/main/java/com/braze/push/BrazeNotificationStyleFactory.kt
@@ -11,7 +11,6 @@ import android.os.Build
 import android.os.Bundle
 import android.view.View
 import android.widget.RemoteViews
-import androidx.annotation.RequiresApi
 import androidx.annotation.VisibleForTesting
 import androidx.core.app.NotificationCompat
 import com.braze.Braze
@@ -92,7 +91,7 @@ open class BrazeNotificationStyleFactory {
                 style = getConversationalPushStyle(notificationBuilder, payload)
             } else if (payload.bigImageUrl != null) {
                 style =
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && payload.isInlineImagePush) {
+                    if (payload.isInlineImagePush) {
                         brazelog { "Rendering push notification with custom inline image style" }
                         getInlineImageStyle(payload, notificationBuilder)
                     } else {
@@ -204,7 +203,6 @@ open class BrazeNotificationStyleFactory {
          * null style is returned.
          */
         @Suppress("LongMethod", "ReturnCount")
-        @RequiresApi(api = Build.VERSION_CODES.M)
         fun getInlineImageStyle(
             payload: BrazeNotificationPayload,
             notificationBuilder: NotificationCompat.Builder