+++ android-sdk-ui/src/main/java/com/braze/push/support/HtmlUtils.kt
@@ -2,7 +2,6 @@
 
 package com.braze.push.support
 
-import android.os.Build
 import android.text.Html
 import com.braze.configuration.BrazeConfigurationProvider
 import com.braze.support.BrazeLogger.brazelog
@@ -24,11 +23,7 @@ fun String.getHtmlSpannedTextIfEnabled(
         return this
     }
     return if (configurationProvider.isPushNotificationHtmlRenderingEnabled) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            Html.fromHtml(this, Html.FROM_HTML_MODE_LEGACY)
-        } else {
-            Html.fromHtml(this)
-        }
+        Html.fromHtml(this, Html.FROM_HTML_MODE_LEGACY)
     } else {
         this
     }