+++ droidboy/src/main/java/com/appboy/sample/activity/settings/SettingsFragment.kt
@@ -240,15 +240,11 @@ class SettingsFragment : PreferenceFragmentCompat() {
             }
         }
         setClickPreference("location_runtime_permission_dialog") {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                RuntimePermissionUtils.requestLocationPermission(
-                    activity as Activity,
-                    Manifest.permission.ACCESS_FINE_LOCATION,
-                    requestPermissionLauncher
-                )
-            } else {
-                showToast("Below Android M there is no need to check for runtime permissions.")
-            }
+            RuntimePermissionUtils.requestLocationPermission(
+                activity as Activity,
+                Manifest.permission.ACCESS_FINE_LOCATION,
+                requestPermissionLauncher
+            )
         }
     }